Finished but untested Lab Sampling endpoint

+++ models/sensor_reading.go
@@ -4,7 +4,7 @@ import "github.com/jinzhu/gorm"
 
 type SensorReading struct {
 	gorm.Model
-	SensorID int
-	Value    float32
-	SampleID int
+	SensorID uint
+	Value    float64
+	SampleID uint
 }
